# server/dist/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "Partial_User-or-null_": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "bvn_hash": {
                        "type": "string"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "description": "Make all properties in T optional"
            },
            "Partial_User_": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "bvn_hash": {
                        "type": "string"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "description": "Make all properties in T optional"
            },
            "UpdateUserDto": {
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
            "ErrorResponseDto": {
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid request parameters"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "double",
                        "description": "HTTP status code",
                        "example": 400
                    },
                    "name": {
                        "type": "string",
                        "description": "Error name/type",
                        "example": "ValidationError"
                    }
                },
                "required": [
                    "message",
                    "statusCode",
                    "name"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "bvn_hash": {
                        "type": "string"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "email"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token obtained from /auth/login"
            }
        }
    },
    "info": {
        "title": "coown-api",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Salawu Olanrewaju Joseph"
        }
    },
    "paths": {
        "/users/{id}": {
            "get": {
                "operationId": "GetUserById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Partial_User-or-null_"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "id": "550e8400-e29b-41d4-a716-446655440000",
                                            "firstName": "John",
                                            "lastName": "Doe",
                                            "email": "john.doe@example.com",
                                            "phone": "+2348012345678",
                                            "verified": true,
                                            "role": "user",
                                            "isActive": true,
                                            "created_at": "2024-01-15T10:30:00.000Z",
                                            "updatedAt": "2024-01-15T10:30:00.000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "summary": "Get user by ID",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Partial_User_"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "firstName": "Jane",
                                            "lastName": "Doe",
                                            "phone": "+2348012345678"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "description": "UPDATE USER",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": "550e8400-e29b-41d4-a716-446655440000"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "description": "DELETE USER",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "operationId": "ListAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Partial_User_"
                                    },
                                    "type": "array"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": [
                                            {
                                                "id": "550e8400-e29b-41d4-a716-446655440000",
                                                "firstName": "John",
                                                "lastName": "Doe",
                                                "email": "john.doe@example.com",
                                                "role": "user",
                                                "verified": true,
                                                "isActive": true,
                                                "created_at": "2024-01-15T10:30:00.000Z",
                                                "updatedAt": "2024-01-15T10:30:00.000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "LIST ALL USERS",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "role",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "isActive",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ]
            }
        },
...
        "/pools/{id}/users": {
            "get": {
                "operationId": "GetPoolUsers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pool Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Pool not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Pools"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}